Update S3_mounted_on_EC2.sh
We need a bucket name and iam role from a user, but I don't think that the user has to specify the name of a folder, that is going to be used as a mount point. So I've changed it to /images, we are going to need this folder in a php script too.

Also installed s3fs form Installation from pre-built packages from epel repo instead of compiling and installing.